renamed parser module to heading
