+++ src/chapter.c
@@ -6,7 +6,7 @@
 #include <unistd.h>
 
 #include "chapter.h"
-#include "parser.h"
+#include "heading.h"
 #include "pretty_printer.h"
 #include "stream.h"
 #include "string_util.h"
